refactor(components): refactor badge (#3366)
