+++ packages/components/badge/__tests__/badge.spec.ts
@@ -1,5 +1,5 @@
 import { mount } from '@vue/test-utils'
-import Badge from '../src/index.vue'
+import Badge from '../src/badge.vue'
 
 const AXIOM = 'Rem is the best girl'
 
@@ -8,7 +8,7 @@ describe('Badge', () => {
     const wrapper = mount(Badge, {
       props: { value: 80 },
     })
-    expect(wrapper.vm.content).toEqual(80)
+    expect(wrapper.vm.content).toEqual('80')
   })
 
   test('is fixed', () => {
@@ -46,6 +46,6 @@ describe('Badge', () => {
     })
     expect(wrapper.vm.content).toEqual('100+')
     await wrapper.setProps({ value: 80 })
-    expect(wrapper.vm.content).toEqual(80)
+    expect(wrapper.vm.content).toEqual('80')
   })
 })
